kwork_parser_bot/requirements.txt
aiogram==3.0.0b7
python-dotenv~=1.0.0
pydantic~=1.10.12
loguru~=0.7.0
jinja2~=3.1.2
stringcase~=1.2.0
apscheduler~=3.10.1
sqlalchemy==2.0.19
asyncpg
psycopg2
black
aiohttp~=3.8.5
redis~=4.6.0
redis-om
fastapi~=0.101.0
pytz~=2023.3
uvicorn~=0.23.2
prisma
passlib